Move Identity DI into Platform lib.
