+++ src/web/BL_Python/web/middleware/dependency_injection.py
@@ -1,8 +1,6 @@
 from functools import partial
 from typing import Any, Protocol, Tuple, cast
 
-from BL_Python.platform.identity.user_loader import Loader
-from BL_Python.platform.identity.user_loader import T as TUserMixin
 from BL_Python.programming.patterns.dependency_injection import LoggerModule
 from connexion import ConnexionMiddleware, FlaskApp
 from connexion.apps.flask import FlaskASGIApp, FlaskOperation
@@ -21,19 +19,6 @@ class MiddlewareRoutine(Protocol):
     def __call__(
         self, scope: Scope, receive: Receive, send: Send, *args: Any
     ) -> None: ...
-
-
-class IdentityModule(Module):
-    def __init__(
-        self,
-        loader: Loader[TUserMixin],
-    ) -> None:
-        super().__init__()
-        self._loader = loader
-
-    @override
-    def configure(self, binder: Binder) -> None:
-        binder.bind(Loader, to=self._loader)
 
 
 class AppModule(Module):
